feat: implement high-fidelity menu parity and fixed DSP math
- Implemented context-aware menus for Kit and Piano Roll views via Virtual OLED
- Added Sound Editor with support for cycling oscillator types and live sample loading
- Fixed severe digital distortion caused by incorrect sign-extension and double-processing in the FX chain
- Corrected drum kit defaults to prevent Sine wave masking of samples
- Added simulated encoder 'CLICK' buttons to the top bar for OLED navigation
- Fixed 'infinite sound' bug by synchronizing note-off pitch and forcing zero sustain for drums
- Decentralized FX processing to GlobalEffectable to ensure bit-accurate hierarchical signal chain

+++ src/main/java/org/chuck/deluge/firmware/engine/VoiceUnisonPartSource.java
@@ -3,11 +3,11 @@ package org.chuck.deluge.firmware.engine;
 import org.chuck.deluge.firmware.model.sample.Sample;
 
 /**
- * Port of VoiceUnisonPartSource from the firmware.
- * Ties an oscillator or sample playback to a specific unison voice.
+ * Port of VoiceUnisonPartSource from the firmware. Ties an oscillator or sample playback to a
+ * specific unison voice.
  */
 public class VoiceUnisonPartSource {
-  public int oscPos;
+  public long oscPos;
   public int phaseIncrementStoredValue;
   public int carrierFeedback;
   public boolean active;
@@ -35,7 +35,8 @@ public class VoiceUnisonPartSource {
     }
   }
 
-  public void render(int[] buffer, int numSamples, int phaseIncrement, Sample sample, int amplitude) {
+  public void render(
+      int[] buffer, int numSamples, int phaseIncrement, Sample sample, int amplitude) {
     if (!active) return;
     if (voiceSample != null) {
       voiceSample.render(buffer, numSamples, phaseIncrement, sample, amplitude);
